+++ vizro-ai/tests/unit/vizro-ai/dashboard/conftest.py
@@ -1,26 +1,6 @@
-from typing import Any
-
 import pytest
 import vizro.models as vm
-from langchain.output_parsers import PydanticOutputParser
-from langchain_community.llms.fake import FakeListLLM
 from langchain_core.messages import HumanMessage
-
-
-class MockStructuredOutputLLM(FakeListLLM):
-    def bind_tools(self, tools: list[Any]):
-        return super().bind(tools=tools)
-
-    def with_structured_output(self, schema, **kwargs):
-        llm = self
-        output_parser = PydanticOutputParser(pydantic_object=schema)
-        return llm | output_parser
-
-
-@pytest.fixture
-def fake_llm():
-    response = ['{"text":"this is a card","href":""}']
-    return MockStructuredOutputLLM(responses=response)
 
 
 @pytest.fixture
@@ -50,9 +30,3 @@ def message_output_error():
         "df_info": None,
         "validation_error": "ValidationError",
     }
-
-
-@pytest.fixture
-def fake_llm_invalid():
-    response = ['{"text":"this is a card", "href": "", "icon": "summary"}']
-    return MockStructuredOutputLLM(responses=response)
